Fix dropbox size decimal

+++ kDrive/UI/Controller/Files/DropBox/ManageDropBoxViewController.swift
@@ -322,7 +322,7 @@ extension ManageDropBoxViewController: FooterButtonDelegate {
         let password = getSetting(for: .optionPassword) ? (getValue(for: .optionPassword) as? String) : ""
         let validUntil = getSetting(for: .optionDate) ? (getValue(for: .optionDate) as? Date) : nil
         let limitFileSize: BinarySize?
-        if getSetting(for: .optionSize), let size = getValue(for: .optionSize) as? Int {
+        if getSetting(for: .optionSize), let size = getValue(for: .optionSize) as? Double {
             limitFileSize = .gigabytes(size)
         } else {
             limitFileSize = nil
